+++ progroup/users/routes.py
@@ -9,12 +9,22 @@ users = Blueprint('users', __name__)
 
 @users.route("/get_users")
 def get_users():
+
+      # Check the user is logged in
+    if 'user' not in session:
+        return redirect(url_for("authentication.login"))
+
     users = list(mongo.db.users.find())
     return render_template("users/users.html", users=users)
 
 
 @users.route("/add_user", methods=["GET", "POST"])
 def add_user():
+
+  # Check the user is logged in
+    if 'user' not in session:
+        return redirect(url_for("authentication.login"))
+
     if request.method == "POST":
         is_admin = "admin" if request.form.get("is_admin") else "user" 
         users = {
@@ -37,6 +47,11 @@ def add_user():
 
 @users.route("/edit_user/<user_id>", methods=["GET", "POST"])
 def edit_user(user_id):
+
+  # Check the user is logged in
+    if 'user' not in session:
+        return redirect(url_for("authentication.login"))
+
     if request.method == "POST":
         is_admin = "admin" if request.form.get("is_admin") else "user" 
         updated_user = {"$set": 
@@ -60,6 +75,11 @@ def edit_user(user_id):
 
 @users.route('/delete_user/<user_id>')
 def delete_user(user_id):
+
+  # Check the user is logged in
+    if 'user' not in session:
+        return redirect(url_for("authentication.login"))
+
     mongo.db.users.delete_one({"_id": ObjectId(user_id)})
     flash("User Deleted")
     return redirect(url_for("users.get_users"))
